Moving convection model determination code into new cnvmodel library from map_addmodel and solve_model binaries
