--- codebase/superdarn/src.bin/tk/tool/map_addmodel.1.14/map_addmodel.h
@@ -1,9 +1,0 @@
-
-/* statistical models */
-
-#define RG96  1
-#define PSR10 2
-#define CS10  4
-#define TS18  8
-#define TS18_Kp 16
-
