+++ codebase/superdarn/src.lib/tk/cnvmodel.1.0/src/makefile
@@ -1,0 +1,15 @@
+# Makefile for cnvmodel library
+# =============================
+# by E.G.Thomas
+#
+#
+include $(MAKECFG).$(SYSTEM)
+INCLUDE=-I$(IPATH)/base -I$(IPATH)/general -I$(IPATH)/superdarn
+SRC = cnvmodel.c
+OBJS = cnvmodel.o
+INC=$(IPATH)/superdarn
+LINK="1"
+DSTPATH=$(LIBPATH)
+OUTPUT = cnvmodel
+
+include $(MAKELIB).$(SYSTEM)
